Allow sudo for all instead of just members of the wheel group
We already claim this in the README:

> This being an image for development use, we have installed sudo and
  configured all users in the container to use it without providing a
  password.

Previously we did this by making sure the typical users were in the
wheel group.

But with a recent update (probably to the underlying alpine image --
but I'm not sure) the process doesn't have knowledge of this unless we
run Docker with the `--group-add wheel` option.

We want to avoid that. So let's just implement what we actually claim
in the documentation. This should also be better for the non-standard
use case where someone runs the container using a "non-stanard" user.